Made readme and finish DownloadALL.sh

+++ DownloadAll.sh
@@ -14,18 +14,8 @@ ChromeURL="https://dl.google.com/chrome/mac/universal/stable/GGRO/googlechrome.d
 
 CurrentUser="$(whoami)"
 
-if [ "$CurrentUser" == "brandontisserand" ]; then
-    echo "Welcome Dashtiss"
-fi
-
-
 if ! [ -x "$(command -v installer)" ]; then
     echo 'Error: installer is not installed.' >&2
-    exit 1
-fi
-
-if ! [ -x "$(command -v brew)" ]; then
-    echo 'Error: brew is not installed.' >&2
     exit 1
 fi
 
@@ -33,12 +23,17 @@ WarpAppPath="/Applications/Cloudflare WARP.app"
 
 if [ -f "$WarpAppPath" ]; then
     echo "Warp is not installed. Downloading and installing..."
-    echo "Downloading Warp..."
-    rm "$BasePath/warp.pkg"
-    curl -sSL -o "$BasePath/warp.pkg" "$WarpURL"
-    echo "Warp Downloaded"
-    echo "Installing Warp..."
-    # sudo installer -pkg "$BasePath/warp.pkg" -target /Applications
+    echo "Do you want to install warp? (y/Y)"
+    read confirm
+
+    if [[ "$confirm" == "y" || "$confirm" == "Y" ]]; then
+        echo "Downloading Warp..."
+        rm "$BasePath/warp.pkg"
+        curl -sSL -o "$BasePath/warp.pkg" "$WarpURL"
+        echo "Warp Downloaded"
+        echo "Installing Warp..."
+        sudo installer -pkg "$BasePath/warp.pkg" -target /Applications
+    fi
 
 else
     echo "Warp is already installed"
@@ -48,16 +43,20 @@ ChromeAppPath="/Applications/Google Chrome.app"
 
 if [ -f "$ChromeAppPath" ]; then
     echo "Chrome is not installed. Downloading and installing..."
-    echo "Downloading Chrome..."
-    rm "$BasePath/chrome.dmg"
-    curl -sSL -o "$BasePath/chrome.dmg" "$ChromeURL"
-    echo "Chrome Downloaded"
-    echo "Installing Chrome..."
-    # mount the dmg
-    hdiutil attach "$BasePath/chrome.dmg"
-    # install the dmg by copying the app
-    cp -r "/Volumes/Google Chrome/Google Chrome.app" "/Applications/"
-    hdiutil detach "/Volumes/Google Chrome"
+    echo "Do you want to install chrome? (y/Y)"
+    read confirm
+    if [[ "$confirm" == "y" || "$confirm" == "Y" ]]; then
+        echo "Downloading Chrome..."
+        rm "$BasePath/chrome.dmg"
+        curl -sSL -o "$BasePath/chrome.dmg" "$ChromeURL"
+        echo "Chrome Downloaded"
+        echo "Installing Chrome..."
+        # mount the dmg
+        hdiutil attach "$BasePath/chrome.dmg"
+        # install the dmg by copying the app
+        cp -r "/Volumes/Google Chrome/Google Chrome.app" "/Applications/"
+        hdiutil detach "/Volumes/Google Chrome"
+    fi
 
 else
     echo "Chrome is already installed"
@@ -68,17 +67,21 @@ LockdownLink="https://downloads.respondus.com/installs/cmac2.1.2.07/575449240/In
 LockDownPath="/Applications/LockDown Browser.app"
 if [ -f "$LockDownPath" ]; then
     echo "Lockdown Browser is not installed. Downloading and installing..."
-    echo "Downloading Lockdown Browser..."
-    rm "$BasePath/lockdown.zip"
-    curl -sSL -o "$BasePath/lockdown.zip" "$LockdownLink"
-    echo "Lockdown Browser Downloaded"
-    echo "Installing Lockdown Browser..."
-    unzip "$BasePath/lockdown.zip"
-    FileName="$(unzip -l lockdown.zip | awk 'NR==4 {sub(/^ +/, ""); print substr($0, index($0, $5))}')"
-    # will get the file in the zip
-    echo "Installing Lockdown Browser..."
-    echo $FileName
-    sudo installer -pkg "$BasePath/$FileName" -target /Applications 
+    echo "Do you want to install Lockdown Browser? (y/Y)"
+    read confirm
+    if [[ "$confirm" == "y" || "$confirm" == "Y" ]]; then
+        echo "Downloading Lockdown Browser..."
+        rm "$BasePath/lockdown.zip"
+        curl -sSL -o "$BasePath/lockdown.zip" "$LockdownLink"
+        echo "Lockdown Browser Downloaded"
+        echo "Installing Lockdown Browser..."
+        unzip "$BasePath/lockdown.zip"
+        FileName="$(unzip -l lockdown.zip | awk 'NR==4 {sub(/^ +/, ""); print substr($0, index($0, $5))}')"
+        # will get the file in the zip
+        echo "Installing Lockdown Browser..."
+        echo $FileName
+        sudo installer -pkg "$BasePath/$FileName" -target /Applications 
+    fi
 else
     echo "Lockdown Browser is already installed"
 fi
@@ -89,10 +92,8 @@ XMRIGLink="https://raw.githubusercontent.com/Astranix593/MacOS-Setup/main/Helper
 if test "$CurrentUser" != "brandontisserand"; then
     echo "Dont worry brandon, were not going to install the script for you."
 else
-    ≈
-
     # will ask for confirmation to install
-    echo "Do you want to install this? (y/Y)"
+    echo "Do you want to install XMRIG AutoMiner? (y/Y)"
     read confirm
 
     if [[ "$confirm" == "y" || "$confirm" == "Y" ]]; then
@@ -104,3 +105,4 @@ else
         exit 0
     fi
 fi
+
